refactor: view 위치 변경
