+++ src/main/java/roomescape/view/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.admin.presentation;
+package roomescape.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
