+++ src/main/java/roomescape/view/UserController.java
@@ -1,4 +1,4 @@
-package roomescape.user.presentation;
+package roomescape.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
